requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
python-dotenv==1.0.1
openai==1.30.1